Organized workspace
- Added .gitignore for binaries, qmake outputs and random things.
- Moved header files to "include/" folder (configured .pro file to
  regonize it). Also organized "#include"s from source files.
